+++ composeApp/src/commonMain/kotlin/world/hachimi/app/di/di.kt
@@ -2,7 +2,25 @@ package world.hachimi.app.di
 
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.viewModelOf
-import world.hachimi.app.model.*
+import world.hachimi.app.model.ArtworkDetailViewModel
+import world.hachimi.app.model.AuthViewModel
+import world.hachimi.app.model.CategorySongsViewModel
+import world.hachimi.app.model.ForgetPasswordViewModel
+import world.hachimi.app.model.HomeViewModel
+import world.hachimi.app.model.MyPRViewModel
+import world.hachimi.app.model.PlaylistDetailViewModel
+import world.hachimi.app.model.PlaylistViewModel
+import world.hachimi.app.model.PublicPlaylistViewModel
+import world.hachimi.app.model.PublishViewModel
+import world.hachimi.app.model.PublishedTabViewModel
+import world.hachimi.app.model.RecentPlayViewModel
+import world.hachimi.app.model.RecentPublishViewModel
+import world.hachimi.app.model.RecommendViewModel
+import world.hachimi.app.model.ReviewDetailViewModel
+import world.hachimi.app.model.ReviewViewModel
+import world.hachimi.app.model.SearchViewModel
+import world.hachimi.app.model.UserSpaceViewModel
+import world.hachimi.app.model.WeeklyHotViewModel
 
 fun Module.applyViewModels() {
     viewModelOf(::RecentPublishViewModel)
@@ -23,4 +41,5 @@ fun Module.applyViewModels() {
     viewModelOf(::CategorySongsViewModel)
     viewModelOf(::PublishedTabViewModel)
     viewModelOf(::ArtworkDetailViewModel)
+    viewModelOf(::PublicPlaylistViewModel)
 }