+++ composeApp/src/commonMain/kotlin/world/hachimi/app/nav/Route.kt
@@ -25,6 +25,9 @@ sealed class Route {
             data object List: MyPlaylist()
             data class Detail(val playlistId: Long): MyPlaylist()
         }
+
+        data class PublicPlaylist(val playlistId: Long): Root()
+
         sealed class CreationCenter: Root() {
             companion object Companion {
                 val Default = MyArtwork