+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/playlist/PublicPlaylistScreen.kt
@@ -1,7 +1,7 @@
 package world.hachimi.app.ui.playlist
 
+import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -21,7 +21,6 @@ import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
@@ -31,9 +30,11 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.painter.ColorPainter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import coil3.compose.LocalPlatformContext
+import coil3.compose.rememberAsyncImagePainter
 import coil3.request.ImageRequest
 import coil3.request.crossfade
 import dev.chrisbanes.haze.hazeSource
@@ -42,33 +43,39 @@ import hachimiworld.composeapp.generated.resources.Res
 import hachimiworld.composeapp.generated.resources.play_all
 import hachimiworld.composeapp.generated.resources.playlist_cover_cd
 import hachimiworld.composeapp.generated.resources.playlist_description_placeholder
-import hachimiworld.composeapp.generated.resources.playlist_private_badge
+import hachimiworld.composeapp.generated.resources.playlist_favorite
 import hachimiworld.composeapp.generated.resources.playlist_song_count
 import hachimiworld.composeapp.generated.resources.playlist_songs_list
+import hachimiworld.composeapp.generated.resources.playlist_unfavorite
 import hachimiworld.composeapp.generated.resources.song_cover_cd
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.koinInject
 import org.koin.compose.viewmodel.koinViewModel
-import world.hachimi.app.api.module.PlaylistModule
+import soup.compose.material.motion.animation.materialFadeThrough
 import world.hachimi.app.model.GlobalStore
 import world.hachimi.app.model.InitializeStatus
-import world.hachimi.app.model.PlaylistDetailViewModel
+import world.hachimi.app.model.PublicPlaylistViewModel
+import world.hachimi.app.nav.Route
 import world.hachimi.app.ui.LocalContentInsets
 import world.hachimi.app.ui.component.LoadingPage
 import world.hachimi.app.ui.component.ReloadPage
+import world.hachimi.app.ui.design.components.AccentButton
 import world.hachimi.app.ui.design.components.Button
 import world.hachimi.app.ui.design.components.Icon
+import world.hachimi.app.ui.design.components.LinearProgressIndicator
 import world.hachimi.app.ui.design.components.LocalContentColor
-import world.hachimi.app.ui.design.components.TagBadge
 import world.hachimi.app.ui.design.components.Text
-import world.hachimi.app.ui.playlist.components.EditDialog
+import world.hachimi.app.ui.player.fullscreen.components.UserChip
 import world.hachimi.app.ui.playlist.components.SongItem
+import world.hachimi.app.ui.theme.PreviewTheme
+import world.hachimi.app.util.fillMaxWidthIn
 import kotlin.time.Duration.Companion.seconds
 
 @Composable
-fun PlaylistDetailScreen(
+fun PublicPlaylistScreen(
     playlistId: Long,
-    vm: PlaylistDetailViewModel = koinViewModel(),
+    vm: PublicPlaylistViewModel = koinViewModel(),
+    global: GlobalStore = koinInject()
 ) {
     DisposableEffect(vm, playlistId) {
         vm.mounted(playlistId)
@@ -77,46 +84,49 @@ fun PlaylistDetailScreen(
         }
     }
 
-    val global = koinInject<GlobalStore>()
-    when (vm.initStatus) {
-        InitializeStatus.INIT -> LoadingPage()
-        InitializeStatus.FAILED -> ReloadPage(onReloadClick = { vm.retry() })
-        InitializeStatus.LOADED -> Box(Modifier.fillMaxSize()) {
-            vm.playlistInfo?.let { info ->
-                LazyColumn(
+    AnimatedContent(
+        targetState = vm.initStatus,
+        transitionSpec = { materialFadeThrough() }
+    ) { initStatus ->
+        when (initStatus) {
+            InitializeStatus.INIT -> LoadingPage()
+            InitializeStatus.FAILED -> ReloadPage(onReloadClick = { vm.retry() })
+            InitializeStatus.LOADED -> Box(Modifier.fillMaxSize()) {
+                val playlistInfo = vm.playlistInfo
+                val userInfo = vm.creatorProfile
+                if (playlistInfo != null && userInfo != null) LazyColumn(
                     modifier = Modifier.fillMaxSize(),
                     contentPadding = PaddingValues(24.dp),
                     verticalArrangement = Arrangement.spacedBy(16.dp)
                 ) {
                     item {
-                        Header(vm, info, modifier = Modifier.fillMaxWidth())
+                        Header(
+                            modifier = Modifier.fillMaxWidthIn(),
+                            username = userInfo.username,
+                            avatarUrl = userInfo.avatarUrl,
+                            description = playlistInfo.description,
+                            title = playlistInfo.name,
+                            coverUrl = playlistInfo.coverUrl,
+                            count = vm.songs.size,
+                            onNavToUserClick = { global.nav.push(Route.Root.PublicUserSpace(userInfo.uid)) },
+                            onPlayAllClick = { vm.playAll() },
+                            isFavorite = vm.isFavorite,
+                            operating = vm.operating,
+                            onFavoriteClick = { vm.favorite(it) }
+                        )
                     }
 
                     itemsIndexed(vm.songs, key = { _, item -> item.songId }) { index, song ->
                         SongItem(
-                            modifier = Modifier.fillMaxWidth(),
+                            modifier = Modifier.fillMaxWidthIn(),
                             orderIndex = index,
                             title = song.title,
-                            onClick = {
-                                global.player.insertToQueue(
-                                    GlobalStore.MusicQueueItem(
-                                        id = song.songId,
-                                        displayId = song.songDisplayId,
-                                        name = song.title,
-                                        artist = song.uploaderName,
-                                        duration = song.durationSeconds.seconds,
-                                        coverUrl = song.coverUrl,
-                                        explicit = null // TODO(playlist): Get explicit info
-                                    ), true, false
-                                )
-                            },
+                            onClick = { vm.play(song) },
                             coverUrl = song.coverUrl,
                             artist = song.uploaderName,
                             duration = song.durationSeconds.seconds,
-                            editable = true,
-                            onRemoveClick = {
-                                vm.removeFromPlaylist(song.songId)
-                            }
+                            editable = false,
+                            onRemoveClick = {}
                         )
                     }
 
@@ -127,30 +137,34 @@ fun PlaylistDetailScreen(
                         )
                     }
                 }
+                if (vm.loading) LinearProgressIndicator(
+                    Modifier.align(Alignment.TopStart).fillMaxWidthIn()
+                )
             }
-
-            if (vm.loading) CircularProgressIndicator(Modifier.align(Alignment.Center))
-
-            EditDialog(vm)
         }
     }
 }
 
 @Composable
 private fun Header(
-    vm: PlaylistDetailViewModel,
-    info: PlaylistModule.PlaylistItem,
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    title: String,
+    description: String?,
+    coverUrl: String?,
+    username: String,
+    avatarUrl: String?,
+    count: Int,
+    onPlayAllClick: () -> Unit,
+    onNavToUserClick: () -> Unit,
+    isFavorite: Boolean?,
+    operating: Boolean,
+    onFavoriteClick: (Boolean) -> Unit
 ) {
     Column(modifier) {
         Row(Modifier.fillMaxWidth()) {
             Box(
                 modifier = Modifier.size(128.dp)
                     .clip(RoundedCornerShape(24.dp))
-                    .clickable(
-                        onClick = { vm.editCover() },
-                        enabled = !vm.coverUploading
-                    )
                     .background(LocalContentColor.current.copy(0.12f)),
                 contentAlignment = Alignment.Center
             ) {
@@ -158,41 +172,47 @@ private fun Header(
                 AsyncImage(
                     modifier = Modifier.hazeSource(hazeState).fillMaxSize(),
                     model = ImageRequest.Builder(LocalPlatformContext.current)
-                        .data(info.coverUrl)
+                        .data(coverUrl)
                         .crossfade(true)
                         .build(),
                     contentDescription = stringResource(Res.string.playlist_cover_cd),
                     contentScale = ContentScale.Crop,
                     placeholder = ColorPainter(LocalContentColor.current.copy(alpha = 0.12f))
                 )
-                if (vm.coverUploading) {
-                    if (vm.coverUploadingProgress == 0f || vm.coverUploadingProgress == 1f) CircularProgressIndicator()
-                    else CircularProgressIndicator(progress = { vm.coverUploadingProgress })
-                }
-                if (!info.isPublic) {
-                    TagBadge(
-                        hazeState,
-                        tag = stringResource(Res.string.playlist_private_badge),
-                        modifier = Modifier.align(Alignment.BottomEnd).padding(8.dp)
-                    )
-                }
             }
 
             Column(
                 modifier = Modifier.weight(1f).padding(start = 24.dp).height(120.dp)
                     .clip(MaterialTheme.shapes.small)
-                    .clickable(onClick = { vm.edit() })
             ) {
                 Text(
                     modifier = Modifier.fillMaxWidth(),
-                    text = info.name,
+                    text = title,
                     style = MaterialTheme.typography.titleMedium
                 )
                 Text(
                     modifier = Modifier.fillMaxWidth().padding(top = 8.dp),
-                    text = info.description ?: stringResource(Res.string.playlist_description_placeholder),
+                    text = description
+                        ?: stringResource(Res.string.playlist_description_placeholder),
                     style = MaterialTheme.typography.bodySmall,
                     overflow = TextOverflow.Ellipsis
+                )
+
+
+                val avatar = avatarUrl?.let {
+                    rememberAsyncImagePainter(
+                        ImageRequest.Builder(LocalPlatformContext.current)
+                            .crossfade(true)
+                            .data(it)
+                            .build()
+                    )
+                }
+
+                UserChip(
+                    modifier = Modifier.padding(top = 8.dp),
+                    onClick = onNavToUserClick,
+                    avatar = avatar,
+                    name = username
                 )
             }
         }
@@ -203,24 +223,62 @@ private fun Header(
             verticalAlignment = Alignment.CenterVertically
         ) {
             Text(
-                text = stringResource(Res.string.playlist_songs_list), style = MaterialTheme.typography.titleLarge
+                text = stringResource(Res.string.playlist_songs_list),
+                style = MaterialTheme.typography.titleLarge
             )
 
             Text(
                 modifier = Modifier.padding(start = 16.dp),
-                text = stringResource(Res.string.playlist_song_count, vm.songs.size),
+                text = stringResource(Res.string.playlist_song_count, count),
                 style = MaterialTheme.typography.bodySmall
             )
 
             Spacer(Modifier.weight(1f))
             Button(
                 modifier = Modifier,
-                onClick = { vm.playAll() }
+                onClick = onPlayAllClick
             ) {
-                Icon(Icons.Default.PlayArrow, contentDescription = stringResource(Res.string.song_cover_cd))
+                Icon(
+                    Icons.Default.PlayArrow,
+                    contentDescription = stringResource(Res.string.song_cover_cd)
+                )
                 Spacer(Modifier.width(16.dp))
                 Text(stringResource(Res.string.play_all))
             }
+
+            isFavorite?.let { isFavorite ->
+                if (isFavorite) Button(onClick = {
+                    onFavoriteClick(!isFavorite)
+                }, enabled = !operating) {
+                    Text(stringResource(Res.string.playlist_unfavorite))
+                } else AccentButton(onClick = {
+                    onFavoriteClick(!isFavorite)
+                }, enabled = !operating) {
+                    Text(stringResource(Res.string.playlist_favorite))
+                }
+            }
         }
+    }
+}
+
+
+@Composable
+@Preview
+private fun PreviewHeader() {
+    PreviewTheme(background = true) {
+        Header(
+            modifier = Modifier.fillMaxWidthIn(),
+            title = "Title",
+            description = "Description",
+            coverUrl = null,
+            username = "username",
+            avatarUrl = null,
+            count = 10,
+            onPlayAllClick = {},
+            onNavToUserClick = {},
+            isFavorite = false,
+            onFavoriteClick = {},
+            operating = false
+        )
     }
 }